backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.1
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.11.2
